s3_lakehouse_workshop/cloud_resources/requirements.txt
aws-cdk-lib>=2.110.1
constructs>=10.0.0
boto3